+++ processing/src/main/java/io/druid/query/aggregation/LongMaxAggregatorFactory.java
@@ -22,48 +22,27 @@ package io.druid.query.aggregation;
 import com.fasterxml.jackson.annotation.JacksonInject;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.google.common.base.Preconditions;
 import io.druid.java.util.common.StringUtils;
 import io.druid.math.expr.ExprMacroTable;
-import io.druid.math.expr.Parser;
-import io.druid.segment.BaseLongColumnValueSelector;
 import io.druid.segment.ColumnSelectorFactory;
-import io.druid.segment.ColumnValueSelector;
 
 import java.nio.ByteBuffer;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Objects;
 
 /**
  */
-public class LongMaxAggregatorFactory extends AggregatorFactory
+public class LongMaxAggregatorFactory extends SimpleLongAggregatorFactory
 {
-  private final String name;
-  private final String fieldName;
-  private final String expression;
-  private final ExprMacroTable macroTable;
-
   @JsonCreator
   public LongMaxAggregatorFactory(
       @JsonProperty("name") String name,
-      @JsonProperty("fieldName") final String fieldName,
+      @JsonProperty("fieldName") String fieldName,
       @JsonProperty("expression") String expression,
       @JacksonInject ExprMacroTable macroTable
   )
   {
-    Preconditions.checkNotNull(name, "Must have a valid, non-null aggregator name");
-    Preconditions.checkArgument(
-        fieldName == null ^ expression == null,
-        "Must have a valid, non-null fieldName or expression"
-    );
-
-    this.name = name;
-    this.fieldName = fieldName;
-    this.expression = expression;
-    this.macroTable = macroTable;
+    super(macroTable, name, fieldName, expression);
   }
 
   public LongMaxAggregatorFactory(String name, String fieldName)
@@ -74,30 +53,13 @@ public class LongMaxAggregatorFactory extends AggregatorFactory
   @Override
   public Aggregator factorize(ColumnSelectorFactory metricFactory)
   {
-    return new LongMaxAggregator(getLongColumnSelector(metricFactory));
+    return new LongMaxAggregator(makeColumnValueSelectorWithLongDefault(metricFactory, Long.MIN_VALUE));
   }
 
   @Override
   public BufferAggregator factorizeBuffered(ColumnSelectorFactory metricFactory)
   {
-    return new LongMaxBufferAggregator(getLongColumnSelector(metricFactory));
-  }
-
-  private BaseLongColumnValueSelector getLongColumnSelector(ColumnSelectorFactory metricFactory)
-  {
-    return AggregatorUtil.makeColumnValueSelectorWithLongDefault(
-        metricFactory,
-        macroTable,
-        fieldName,
-        expression,
-        Long.MIN_VALUE
-    );
-  }
-
-  @Override
-  public Comparator getComparator()
-  {
-    return LongMaxAggregator.COMPARATOR;
+    return new LongMaxBufferAggregator(makeColumnValueSelectorWithLongDefault(metricFactory, Long.MIN_VALUE));
   }
 
   @Override
@@ -109,28 +71,7 @@ public class LongMaxAggregatorFactory extends AggregatorFactory
   @Override
   public AggregateCombiner makeAggregateCombiner()
   {
-    return new LongAggregateCombiner()
-    {
-      private long max;
-
-      @Override
-      public void reset(ColumnValueSelector selector)
-      {
-        max = selector.getLong();
-      }
-
-      @Override
-      public void fold(ColumnValueSelector selector)
-      {
-        max = Math.max(max, selector.getLong());
-      }
-
-      @Override
-      public long getLong()
-      {
-        return max;
-      }
-    };
+    return new LongMaxAggregateCombiner();
   }
 
   @Override
@@ -140,31 +81,9 @@ public class LongMaxAggregatorFactory extends AggregatorFactory
   }
 
   @Override
-  public AggregatorFactory getMergingFactory(AggregatorFactory other) throws AggregatorFactoryNotMergeableException
-  {
-    if (other.getName().equals(this.getName()) && this.getClass() == other.getClass()) {
-      return getCombiningFactory();
-    } else {
-      throw new AggregatorFactoryNotMergeableException(this, other);
-    }
-  }
-
-  @Override
   public List<AggregatorFactory> getRequiredColumns()
   {
     return Arrays.<AggregatorFactory>asList(new LongMaxAggregatorFactory(fieldName, fieldName, expression, macroTable));
-  }
-
-  @Override
-  public Object deserialize(Object object)
-  {
-    return object;
-  }
-
-  @Override
-  public Object finalizeComputation(Object object)
-  {
-    return object;
   }
 
   @JsonProperty
@@ -187,14 +106,6 @@ public class LongMaxAggregatorFactory extends AggregatorFactory
   }
 
   @Override
-  public List<String> requiredFields()
-  {
-    return fieldName != null
-           ? Collections.singletonList(fieldName)
-           : Parser.findRequiredBindings(Parser.parse(expression, macroTable));
-  }
-
-  @Override
   public byte[] getCacheKey()
   {
     byte[] fieldNameBytes = StringUtils.toUtf8WithNullToEmpty(fieldName);
@@ -209,18 +120,6 @@ public class LongMaxAggregatorFactory extends AggregatorFactory
   }
 
   @Override
-  public String getTypeName()
-  {
-    return "long";
-  }
-
-  @Override
-  public int getMaxIntermediateSize()
-  {
-    return Long.BYTES;
-  }
-
-  @Override
   public String toString()
   {
     return "LongMaxAggregatorFactory{" +
@@ -228,39 +127,5 @@ public class LongMaxAggregatorFactory extends AggregatorFactory
            ", expression='" + expression + '\'' +
            ", name='" + name + '\'' +
            '}';
-  }
-
-  @Override
-  public boolean equals(Object o)
-  {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-
-    LongMaxAggregatorFactory that = (LongMaxAggregatorFactory) o;
-
-    if (!Objects.equals(fieldName, that.fieldName)) {
-      return false;
-    }
-    if (!Objects.equals(expression, that.expression)) {
-      return false;
-    }
-    if (!Objects.equals(name, that.name)) {
-      return false;
-    }
-
-    return true;
-  }
-
-  @Override
-  public int hashCode()
-  {
-    int result = fieldName != null ? fieldName.hashCode() : 0;
-    result = 31 * result + (expression != null ? expression.hashCode() : 0);
-    result = 31 * result + (name != null ? name.hashCode() : 0);
-    return result;
   }
 }
